+++ module/helios-generique/action/SAEEnvoiHelios.class.php
@@ -2,24 +2,32 @@
 
 class SAEEnvoiHelios extends ActionExecutor
 {
+    /**
+     * @throws UnrecoverableException
+     * @throws NotFoundException
+     * @throws DonneesFormulaireException
+     * @throws Exception
+     */
     public function go()
     {
-        /** @var TmpFolder $tmpFolder */
         $tmpFolder = $this->objectInstancier->getInstance(TmpFolder::class);
         $tmp_folder = $tmpFolder->create();
 
-        $result = false;
         try {
             $result = $this->goThrow($tmp_folder);
-        } catch (Exception $e) {
-            throw $e;
         } finally {
             $tmpFolder->delete($tmp_folder);
         }
         return $result;
     }
 
-    public function goThrow($tmp_folder)
+    /**
+     * @throws UnrecoverableException
+     * @throws NotFoundException
+     * @throws DonneesFormulaireException
+     * @throws Exception
+     */
+    public function goThrow(string $tmp_folder): bool
     {
         $donneesFormulaire = $this->getDonneesFormulaire();
         if (! $donneesFormulaire->get('envoi_signature') && ! $donneesFormulaire->get('fichier_pes_signe')) {
@@ -28,62 +36,28 @@ class SAEEnvoiHelios extends ActionExecutor
             $donneesFormulaire->addFileFromData('fichier_pes_signe', $file_name[0] ?? '', $fichier_pes);
         }
 
-        $pes_aller = $donneesFormulaire->copyFile('fichier_pes_signe', $tmp_folder, 0, "pes_aller");
-        $pes_retour = $donneesFormulaire->copyFile('fichier_reponse', $tmp_folder, 0, "pes_retour");
-
-        if ($donneesFormulaire->get('iparapheur_historique')) {
-            $iparapheur_historique = $donneesFormulaire->copyFile('iparapheur_historique', $tmp_folder, 0, "iparapheur_historique");
-        } else {
-            $iparapheur_historique = false;
-        }
-
-        $transactionsInfo = [
-                'unique_id' => $donneesFormulaire->get('tedetis_transaction_id'),
-                'date' => date("Y-m-d"),
-                'description' => 'inconnu',
-                'pes_retour_description' => 'inconnu',
-                'pes_aller' => $pes_aller,
-                'pes_retour' => $pes_retour,
-                'pes_aller_original_filename' => $donneesFormulaire->getFileName('fichier_pes_signe', 0),
-                'pes_retour_original_filename' => $donneesFormulaire->getFileName('fichier_reponse', 0),
-                'pes_description' => 'inconnu',
-                'pes_aller_content' => $donneesFormulaire->getFileContent('fichier_pes_signe'),
-                'iparapheur_historique' => $iparapheur_historique
-        ];
-
         /** @var SEDAConnecteur $heliosSEDA */
         $heliosSEDA = $this->getConnecteur('Bordereau SEDA');
 
         /** @var SAEConnecteur $sae */
         $sae = $this->getConnecteur('SAE');
 
+        $fluxData = new FluxDataSedaHelios($donneesFormulaire);
+        $bordereau = $heliosSEDA->getBordereauNG($fluxData);
 
-        if ($heliosSEDA instanceof SedaNG) {
-            /** @var SedaNG $heliosSEDA */
-            $fluxData = new FluxDataSedaHelios($donneesFormulaire);
-            $bordereau = $heliosSEDA->getBordereauNG($fluxData);
-            $donneesFormulaire->addFileFromData('sae_bordereau', "bordereau.xml", $bordereau);
-            $transferId = $sae->getTransferId($bordereau);
-            $donneesFormulaire->setData("sae_transfert_id", $transferId);
-
-            try {
-                $heliosSEDA->validateBordereau($bordereau);
-            } catch (Exception $e) {
-                $message = $e->getMessage() . " : <br/><br/>";
-                foreach ($heliosSEDA->getLastValidationError() as $erreur) {
-                    $message .= $erreur->message . "<br/>";
-                }
-                throw new Exception($message);
+        try {
+            $heliosSEDA->validateBordereau($bordereau);
+        } catch (Exception $e) {
+            $message = $e->getMessage() . " : <br/><br/>";
+            foreach ($heliosSEDA->getLastValidationError() as $erreur) {
+                $message .= $erreur->message . "<br/>";
             }
-
-            $archive_path = $tmp_folder . "/archive.tar.gz";
-            // ! generateArchive doit être postérieur à getBordereauNG afin que la liste des fichiers à traiter (file_list de FluxDataSedaDefault) soit renseignée.
-            $heliosSEDA->generateArchive($fluxData, $archive_path);
-        } else {
-            $bordereau = $heliosSEDA->getBordereau($transactionsInfo);
-            $archive_path = $sae->generateArchive($bordereau, $tmp_folder);
+            throw new Exception($message);
         }
 
+        $archive_path = $tmp_folder . "/archive.tar.gz";
+        // ! generateArchive doit être postérieur à getBordereauNG afin que la liste des fichiers à traiter (file_list de FluxDataSedaDefault) soit renseignée.
+        $heliosSEDA->generateArchive($fluxData, $archive_path);
 
         $transferId = $sae->getTransferId($bordereau);
         $donneesFormulaire->setData("sae_transfert_id", $transferId);