+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - Suppression de la classe d'action `FournisseurCommandeEnvoieIparapheur`
 - L'envoi au iparapheur du flux Commande générique est standardisé avec les autres flux
 - Suppression de la plupart des scripts dans batch/ installation/ et script/. Ceux qui restent ne devrait plus être utilisé car ils doivent être remplacés par un script console
+- Suppression de SEDAConnecteur::getBordereau(), remplacé par SEDAConnecteur::getBordereauNG()
 
 
 # [3.1.13] - 2022-04-19